Move test to MainActivityTest
